requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
blinker==1.8.2
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.7
daphne==4.1.2
dataclasses-json==0.6.6
distro==1.9.0
Django==5.0.4
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-timezone-field==6.1.0
django-widget-tweaks==1.5.0
django_paystack==1.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==2.0.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
groq==0.8.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hunter==3.7.0
hyperlink==21.0.0
icecream==2.1.3
idna==3.7
incremental==22.10.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-groq==0.1.4
langchain-text-splitters==0.2.0
langsmith==0.1.67
manhole==1.8.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pip-autoremove==0.10.0
pipdeptree==2.21.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2==2.9.9
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pypaystack2==2.0.3
PySnooper==1.2.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
Twisted==24.3.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zope.interface==6.4
